% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-demo.R
\docType{data}
\name{demo_data}
\alias{demo_data}
\title{Synthetic Dataset for Bootstrap Example}
\format{
A numeric vector of length 20.
}
\usage{
demo_data
}
\description{
A synthetic dataset generated as a sample from a normal distribution.
This dataset is used to demonstrate the bootstrap function in the package.
}
\keyword{datasets}
