% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Dataset}
\format{
A data frame with 15 rows and 2 columns:
\describe{
  \item{DISTANCE}{Numeric, distance from the fire}
  \item{DAMAGE}{Numeric, damage caused by the fire}
}
}
\source{
Manual entry, no specific source.
}
\usage{
fire
}
\description{
This dataset contains information on fire incidents, including the distance from the fire and the damage caused.
}
\keyword{datasets}
