% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_normal.R
\name{mle_normal}
\alias{mle_normal}
\title{Estimate Parameters of Normal Distribution}
\usage{
mle_normal(data)
}
\arguments{
\item{data}{A numeric vector of data points assumed to be drawn from a normal distribution.}
}
\value{
A list containing the maximum likelihood estimates for mean and standard deviation.
}
\description{
This function estimates the mean and standard deviation of a normally
distributed sample using maximum likelihood estimation.
}
\examples{
sample_data <- rnorm(100, mean = 50, sd = 10)
mle_normal(sample_data)
}
