% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Simulate Binomial Experiments}
\usage{
mybin(iter = 100, n = 10, p = 0.7)
}
\arguments{
\item{iter}{The number of iterations to simulate.}

\item{n}{The number of trials per iteration.}

\item{p}{The probability of success on each trial.}
}
\value{
A table of proportions of successes.
}
\description{
This function simulates binomial experiments, where each experiment consists of `n` trials
with success probability `p`, and creates a barplot for the proportions of successes.
}
\examples{
mybin(100, 10, 0.7)

}
