% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{My Boot Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{The number of bootstrap iterations to perform (default is 10000).}

\item{x}{A numeric vector or data from which to sample.}

\item{fun}{The statistic function to apply to each bootstrap sample (default is "mean").}

\item{alpha}{The level of significance to use for the confidence interval (default is 0.05).}

\item{cx}{The expansion factor for the text in the plot (default is 1.5).}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list containing the confidence interval (`ci`) and the point estimate (`pte`) of the statistic.
}
\description{
This function performs bootstrap resampling to estimate the confidence interval of a given statistic.
}
\examples{
set.seed(68)
sample_data <- rnorm(20, mean = 10, sd = 4)
myboot2(x = sample_data)
}
