% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of the normal distribution.}

\item{sigma}{Standard deviation of the normal distribution.}

\item{a}{The x-value to which you want to calculate the probability and shade the area under the curve.}
}
\value{
A list containing the mean (`mu`), standard deviation (`sigma`), and the calculated
probability (`P_X_le_a`).
}
\description{
Calculate and Plot Normal Distribution with Shaded Area
}
\details{
This function plots the normal distribution curve for a given mean (mu) and standard
deviation (sigma), and shades the area under the curve from -infinity to a specified value (a).
It also calculates the probability that a normally distributed random variable is less than
or equal to a (P(X <= a)).
}
\examples{
myncurve(mu = 0, sigma = 1, a = 1)
}
