% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Calculate Optimal Tickets to Sell with Overbooking}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{The total number of seats on the flight.}

\item{gamma}{The probability that the flight will be overbooked.}

\item{p}{The probability of an individual passenger showing up.}
}
\value{
A list containing the optimal number of tickets to sell using the binomial distribution (`nd`), the optimal number using normal approximation (`nc`), and the input parameters `N`, `p`, and `gamma`.
}
\description{
This function calculates the optimal number of tickets to sell based on the probability of overbooking, using both the binomial distribution and the normal approximation. It plots the objective function against the number of tickets sold for both cases and returns a list with the calculated values.
}
\examples{
ntickets(N = 200, gamma = 0.02, p = 0.95)
}
