% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\title{Find Outliers}
\usage{
outliers(I)
}
\arguments{
\item{I}{Numeric vector for which to find outliers.}
}
\value{
A list containing:
\itemize{
  \item{z}{All z-scores of the input vector.}
  \item{outliers}{Z-scores greater than 3.}
  \item{outlier_values}{I values corresponding to z-scores greater than 3.}
  \item{potential_outliers}{Z-scores between 2 and 3.}
  \item{potential_outlier_values}{I values corresponding to z-scores between 2 and 3.}
}
}
\description{
Identifies outliers in a numeric vector based on z-scores. Outliers are defined as observations with z-scores greater than 3, and potential outliers are those with z-scores between 2 and 3.
}
\examples{
outliers(I = c(1, 2, 3, 4, 100))
}
