% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\title{Find Outliers}
\usage{
outliers(I)
}
\arguments{
\item{I}{variable in vector}
}
\value{
z values, z values greater than 3, z values that are possible outliers, I values corresponding to these z values
}
\description{
Find Outliers
}
\examples{
outliers(I = I.df$IRON)
}
